Tests/LibWeb/Text/input/Fetch/fetch-later-abort.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script src="fetch-later-helper.js"></script>
<script>
    promiseTest(async () => {
        try {
            let server = httpTestServer();
            let token = `${Date.now()}-${Math.random().toString(16).slice(2)}`;
            let cookieName = `fetchlater_abort_${token.replace("-", "_")}`;

            let checkUrl = await server.createEcho("GET", `/fetch-later-abort-check-${token}`, {
                status: 200,
                headers: {
                    "Access-Control-Allow-Origin": location.origin,
                    "Access-Control-Allow-Credentials": "true",
                },
                reflect_headers_in_body: true,
            });
            let setUrl = await server.createEcho("POST", `/fetch-later-abort-set-${token}`, {
                status: 204,
                headers: {
                    "Access-Control-Allow-Origin": location.origin,
                    "Access-Control-Allow-Credentials": "true",
                    "Set-Cookie": `${cookieName}=1; Path=/`,
                },
            });

            let controller = new AbortController();
            let result = fetchLater(setUrl, {
                method: "POST",
                credentials: "include",
                body: "defer me",
                activateAfter: 50,
                signal: controller.signal
            });

            // Abort immediately before activation
            controller.abort();
            println(`activated before timeout: ${result.activated}`);

            // Wait enough time to ensure it *would* have activated
            let cookieObserved = await waitForCookie(checkUrl, cookieName, 1000);
            println(`activated after timeout: ${result.activated}`);
            println(`cookie observed: ${cookieObserved}`);
            println(!result.activated && !cookieObserved ? "PASS" : "FAIL");
        } catch (error) {
            println(`FAIL - ${error}`);
        }
    });
</script>
