Tests/LibWeb/Text/expected/Fetch/fetch-later-abort.txt

activated before timeout: false
activated after timeout: false
cookie observed: false
PASS
